split up filters and config from maze_dataset
